- fixed small bug in the handling of Channels API Error callbacks - added sending of links to the cloud - added a connection checker

+++ src/main/java/org/GAEChannel4j/ChannelHandlerAdapter.java
@@ -35,6 +35,6 @@ public class ChannelHandlerAdapter implements IChannelHandler {
     public void close() {
     }
 
-    public void error(String description, int code) {
+    public void error(String description, double code) {
     }
 }
